move support text around
